pyproject.toml
[project]
name = "monitor-control"
version = "0.1.0"
description = "Local monitor control via ddcutil"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.12",
    "rich >=13.7",
    "PyYAML>=6.0",
]

[project.scripts]
monitorctl = "monitor_control.cli:app"

[tool.pytest.ini_options]
testpaths = ["test"]
